latest final draft
I have finished change all of the divs in the accordion to buttons in hopes of improving accessability. I thought that I could simply add focus attributes to each div along with event listeners for 'click' over the element. however this showed to be rather tedious and decided to changed the accordion to divs. This brought another challenge as I quickly learned that there are many tags that don't belong inside a button. I'm realizing that I could have gone a different route and used forms instead for the accordion. either way the app seems to be functional and I'm more concerned over accessibility over transitions.